+++ timeutil/pkg/clock/afterfunc_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/puppetlabs/leg/timeutil/pkg/clock"
 	"github.com/puppetlabs/leg/timeutil/pkg/clock/k8sext"
 	"github.com/stretchr/testify/assert"
-	testclock "k8s.io/apimachinery/pkg/util/clock"
+	testclock "k8s.io/utils/clock/testing"
 )
 
 func TestAfterFunc(t *testing.T) {