+++ timeutil/pkg/clock/k8sext/k8sext.go
@@ -6,23 +6,23 @@ import (
 	"time"
 
 	"github.com/puppetlabs/leg/timeutil/pkg/clock"
-	testclock "k8s.io/apimachinery/pkg/util/clock"
+	k8sclock "k8s.io/utils/clock"
 )
 
-type k8sClock struct{ testclock.Clock }
+type k8sClock struct{ k8sclock.WithTicker }
 
 func (kc *k8sClock) NewTimer(d time.Duration) clock.Timer {
-	t := kc.Clock.NewTimer(d)
-	kc.Clock.Sleep(0)
+	t := kc.WithTicker.NewTimer(d)
+	kc.WithTicker.Sleep(0)
 	return t
 }
 
 func (kc *k8sClock) NewTicker(d time.Duration) clock.Ticker {
-	t := kc.Clock.NewTicker(d)
-	kc.Clock.Sleep(0)
+	t := kc.WithTicker.NewTicker(d)
+	kc.WithTicker.Sleep(0)
 	return t
 }
 
-func NewClock(delegate testclock.Clock) clock.Clock {
+func NewClock(delegate k8sclock.WithTicker) clock.Clock {
 	return &k8sClock{delegate}
 }