+++ timeutil/pkg/clock/timer_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/puppetlabs/leg/timeutil/pkg/clock"
 	"github.com/puppetlabs/leg/timeutil/pkg/clock/k8sext"
 	"github.com/stretchr/testify/assert"
-	testclock "k8s.io/apimachinery/pkg/util/clock"
+	testclock "k8s.io/utils/clock/testing"
 )
 
 func TestTimerCallbackClock(t *testing.T) {